requirements.txt
paho-mqtt
Flask 
gunicorn

